backend/requirements.txt
Django>=4.2,<5
djangorestframework>=3.14,<4
django-cors-headers>=4.3,<5
numpy>=1.24,<3
opencv-python>=4.8,<5
Pillow>=10,<12
